requirements.txt
numpy==2.3.3
pandas==2.3.2
ipython==9.5.0
requests==2.32.5
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0
scikit-learn==1.7.2
nbformat>=4.2.0
kaleido==1.1.0
mlflow>=3.4.0
pyyaml
jupyter
python-dotenv